Separate states extensions by permission to handle states
